AB#635 Fix test for readOnly RichTextEditor contents
