+++ client/tests/webdriver/pages/report/editReport.page.js
@@ -25,41 +25,6 @@ class EditReport extends Page {
     return browser.$(".reportTextField")
   }
 
-  getReportTextContent(selector, multipleElements, index) {
-    const tags = selector.split(" ")
-    selector = tags.map(tag => convertTagForEditor(tag)).join(" ")
-    if (multipleElements) {
-      return this.reportText.$(selector).$$(multipleElements)[index]
-    }
-    return this.reportText.$(selector)
-
-    function convertTagForEditor(tag) {
-      let convertedTag = tag
-      // Some tags are deserialized into different tags having the same appearance
-      switch (tag) {
-        // "b" and "strong" tags are deserialized into "strong" tag
-        case "b":
-          convertedTag = "strong"
-          break
-        // "i" and "em" tags are deserialized into "em" tag
-        case "i":
-          convertedTag = "em"
-          break
-        // "cite" and "blockquote" tags are deserialized into "blockquote" tag
-        case "cite":
-          convertedTag = "blockquote"
-          break
-        // Text content without a tag is wrapped inside a "p" tag
-        case "":
-          convertedTag = "p"
-          break
-        default:
-          return tag
-      }
-      return convertedTag
-    }
-  }
-
   confirmDeleteButton(uuid) {
     return browser.$('//button[text()="Yes, I am sure"]')
   }
